% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_year.R
\name{academic_year}
\alias{academic_year}
\title{Academic Year}
\usage{
academic_year(date_string)
}
\arguments{
\item{date}{A date.}
}
\value{
An integer.
}
\description{
Determine the academic year for a given date
}
\examples{
academic_year("2016-10-31")
}

