% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_year_as_year.R
\name{academic_year_as_year}
\alias{academic_year_as_year}
\title{Academic Year as Year}
\usage{
academic_year_as_year(date_string)
}
\arguments{
\item{date}{A date.}
}
\value{
A date
}
\description{
Determine the academic year for a given date and return it as a Date
}
\examples{
academic_year_as_year("2016-10-31")
}
