% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_to_sp_books.R
\name{link_to_sp_books}
\alias{link_to_sp_books}
\title{HTML link to Scholars Portal Books}
\usage{
link_to_sp_books(ebook_id, anchor_text)
}
\arguments{
\item{ebook_id}{The ID number or string}

\item{anchor_text}{The anchor text}
}
\value{
A string
}
\description{
Given an ID number and some text, make a link to Scholars Portal Books.
}
\examples{
link_to_sp_books("557306", "Marcus Aurelius: A Biography")
}
