% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_to_vufind.R
\name{link_to_vufind}
\alias{link_to_vufind}
\title{HTML link to VuFind record}
\usage{
link_to_vufind(control_number, anchor_text)
}
\arguments{
\item{control_number}{The control number}

\item{anchor_text}{The anchor text}
}
\value{
A string
}
\description{
Given a control number and some text, make a link to VuFind.
}
\examples{
link_to_vufind("a3400583", "R Packages")
}
