% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable_marc245.R
\name{readable_marc245}
\alias{readable_marc245}
\title{Readable MARC 245}
\usage{
readable_marc245(raw_245, include.author = TRUE)
}
\arguments{
\item{title}{A MARC 245 title string (with subfields $a $b $c)}
}
\value{
A string
}
\description{
Convert raw MARC field into something human-readable
}
\examples{
readable_marc245("$a R packages / $c Hadley Wickham.")
}

