% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_of_academic_year.R
\name{start_of_academic_year}
\alias{start_of_academic_year}
\title{Start of Academic Year}
\usage{
start_of_academic_year(given_year)
}
\arguments{
\item{year}{A year.}
}
\value{
A date.
}
\description{
Return the date of the start of a given academic year (always 01 September)
}
\examples{
start_of_academic_year(2018)
}
